fix(ci): GRA-386 (v2) restore main CI at 2cb42ba — rustfmt sweep + Bevy Color baseline (#241)
* fix(ci): GRA-386 (v2) rustfmt sweep + refresh Bevy Color baseline at 2cb42ba

Rebuilt from current origin/main (2cb42ba) after the previous
v1 branch was orphaned by 5 operator-merged commits:
  2cb42ba  fix(astronomy): revert erroneous origin_offset subtraction
  28729b4  perf(porkchop): two-pass refinement
  12197e8  perf(porkchop): class-based cell cap
  d9500a2  wip: working-tree baseline (deleted historical_probes.rs)
  a8ac64f  fix(ui): outer-planet porkchop grid freeze

The new drift:
- cargo fmt --check failed on src/fleets/porkchop.rs,
  src/fleets/visuals.rs, src/ui/fleets_panel.rs,
  src/ui/launch/userdata.rs, src/ui/mod.rs,
  src/ui/transfer_planner.rs. Re-flowed with rustfmt-stable.
- ui-lint audit_bevy_color_literals: 89 NEW Bevy Color literals
  since the last baseline regeneration, all introduced by the
  operator-merged porkchop / astronomy commits (visuals.rs: 11,
  solar_system.rs: 17, system_populator.rs: 7, etc.). Refreshed
  baseline 131 -> 124 entries (some old entries pointed at literals
  the operator-merged commits removed; --emit-baseline reconciles
  the diff).

Verification (local):
- rustfmt --edition 2021 --check   : exit 0
- audit_color32_literals --strict  : exit 0 (13 entries, 0 new)
- audit_bevy_color_literals --strict: exit 0 (124 entries, 0 new)
- audit_b0001                       : exit 0


* fix(test): GRA-386 (v2) add parking_radius_au + body_gm to planner_integration call

The previous v1 PR's fix for this site was on commit d051e0ad (now
orphaned by the v2 reset). The build_grid_for_body_target function
takes 9 arguments (cfg, origin_orbit, dest_orbit, origin_name,
dest_name, category, sim_time_s, parking_radius_au, body_gm) but
the gra_328a_heliocentric_dispatch_yields_a_grid_for_planet_to_planet
test at tests/planner_integration.rs:896 was passing 7. Without this
fix, cargo clippy exits with E0061 and all later gates are skipped.

The (None, GM_SUN) pattern matches the production call site at
src/fleets/porkchop.rs:3851 (the canonical "legacy = system_gm"
fallback used by every test helper in tests/porkchop_rotation_no_snap.rs
and tests/transfer_porkchop.rs). Add the matching use-import for
GM_SUN from helios_ascension::fleets::orbital_mechanics.


* fix(ci): GRA-386 (v3) clippy rust 1.97 lints — len_zero, clone_on_copy, unnecessary_cast, redundant_locals

6 lint fixes surfaced by operator-merged commits on main:
- porkchop.rs: grid.cells.len()>0 → !grid.cells.is_empty() (×2)
- porkchop.rs: planet_orbit.clone() → planet_orbit (×2, KeplerOrbit is Copy)
- visuals.rs: SCALING_FACTOR as f64 → SCALING_FACTOR (f64→f64 cast)
- transfer_planner.rs: redundant let target_entity = target_entity binding


* fix(test): GRA-386 (v4) add parking_radius_au + body_gm to all PorkchopInputs/build_rotating_buffer_for_body_target call sites

v3 only patched tests/planner_integration.rs but missed 14 call sites
of build_rotating_buffer_for_body_target and 2 PorkchopInputs struct
literals in tests/porkchop_rotation_no_snap.rs (60,72,105,115,145,155,
226,251,285,593,674,692,737,832 + line 322) and 1 struct literal in
tests/transfer_porkchop.rs:59.

Clippy rust 1.97 surfaced 15 errors on PR #241 head c49a627:
- 14× E0061 missing args (parking_radius_au: Option<f64>, body_gm: f64)
- 1× E0063 missing fields on PorkchopInputs literal

Per [[feedback-struct-literal-vs-call-site-patching]]: when a Rust struct
grows required fields, grep for BOTH call-site helper args AND direct
struct literals.

Tests default to None + GM_SUN (legacy = system_gm) per existing
planner_integration.rs pattern.


* fix(ci): GRA-386 (v5) refresh Bevy Color baseline after main rebase

Rebased PR #241 onto origin/main (4 new commits since 2cb42ba: porkchop
bilinear fill, 2x persistence camera-denies, Keplerian orbit preview).
The fleets refactor added 3 new Bevy Color literal sites and shifted 108
existing site line numbers in src/fleets/visuals.rs. Without this
emitted baseline the strict audit reports 43 new violations due to line
mismatches alone.


* fix(clippy): remove redundant f64 cast of SCALING_FACTOR

`SCALING_FACTOR` is `pub const f64` (`src/astronomy/systems.rs:18`);
`as f64` is unnecessary. Surfaces after rebase onto main's Keplerian
transfer preview commit.


---------

+++ src/fleets/visuals.rs
@@ -1464,7 +1464,7 @@ mod tests {
         );
 
         // The polyline operates in Bevy render units (AU × SCALING_FACTOR).
-        let scale = SCALING_FACTOR as f64;
+        let scale = SCALING_FACTOR;
 
         // (1) Endpoint must be Jupiter's orbit (apoapsis = r_jupiter_au),
         //     not rotated past it by `mean_motion * elapsed` as the bug did.
@@ -1626,7 +1626,7 @@ mod tests {
         let total_ma_travel = mean_motion * duration_s;
 
         let start_mean_anomaly = transfer_orbit.mean_anomaly_epoch;
-        let scale = SCALING_FACTOR as f64;
+        let scale = SCALING_FACTOR;
 
         // Sample the preview arc — same expression the renderer's
         // `draw_dashed_curve` closure evaluates at fraction `t`.
